Batch Reference Add Many
