+++ src/Weaviate.Client/Models/QueryReference.cs
@@ -3,7 +3,7 @@ namespace Weaviate.Client.Models;
 public record QueryReference
 {
     public string LinkOn { get; init; }
-    public string[] Fields { get; init; }
+    public string[]? Fields { get; init; }
     public MetadataQuery? Metadata { get; init; }
     public IList<QueryReference>? References { get; init; }
 
@@ -15,7 +15,7 @@ public record QueryReference
     )
     {
         LinkOn = linkOn;
-        Fields = fields ?? [];
+        Fields = fields;
         Metadata = metadata;
         References = references;
     }
